apps/cortensor_watcher_bot/requirements.txt
discord.py>=2.3.0
python-telegram-bot>=20.0
requests>=2.31.0
aiohttp>=3.8.0
